single responsibility principle (databasemanager)
