+++ src/main/java/interfaces/IDatabase.java
@@ -5,6 +5,6 @@ import org.json.simple.JSONObject;
 
 public interface IDatabase {
 
-    EventData fetchEvent(String key, Object value);
+    public void initializeDatabase(String fileName);
 
 }
