+++ src/main/java/managers/DatabaseManager.java
@@ -9,34 +9,26 @@ import org.json.simple.parser.ParseException;
 import java.io.FileReader;
 import java.io.IOException;
 
-import data_factories.EventFactory;
-
-import data_objects.EventData;
+import data_factories.EventDataFactory;
 
 
 public class DatabaseManager implements IDatabase {
 
-
+    public final String fileName = "src/main/resources/Database.json";
     public JSONObject fullDatabase;
-    public EventFactory eventFactory;
-
-    /**
-     * Helper function to search JSON arrays
-     */
-    private JSONObject searchJSONArray(JSONArray jsonArray, String key, Object value) {
-        for (Object obj: jsonArray) {
-            JSONObject jsonObject = (JSONObject) obj;
-            if (jsonObject.get(key) == value) {
-                return jsonObject;
-            }
-        }
-        return null;
-    }
+    public EventDataFactory eventFactory;
 
     public DatabaseManager() {
         // Initialize Full Database
+        initializeDatabase(fileName);
+
+        // Initialize Factories
+        this.eventFactory = new EventDataFactory();
+    }
+
+    public void initializeDatabase(String fileName) {
         JSONParser jsonParser = new JSONParser();
-        try (FileReader reader = new FileReader("src/main/resources/Database.json"))
+        try (FileReader reader = new FileReader(fileName))
         {
             //Read JSON file
             Object obj = jsonParser.parse(reader);
@@ -44,19 +36,5 @@ public class DatabaseManager implements IDatabase {
         } catch (ParseException | IOException e) {
             e.printStackTrace();
         }
-
-        // Initialize Factories
-        this.eventFactory = new EventFactory();
     }
-
-    /**
-     * @return EventData with all data from json converted to variables
-     */
-    public EventData fetchEvent(String key, Object value) {
-        JSONArray eventsData = (JSONArray) this.fullDatabase.get("events");
-        JSONObject eventData = searchJSONArray(eventsData, key, value);
-        assert eventData != null;
-        return eventFactory.createEventData(eventData);
-    }
-
 }
