+++ src/main/java/managers/EventDataManager.java
@@ -1,0 +1,32 @@
+package managers;
+
+import data_objects.EventData;
+import org.json.simple.JSONArray;
+import org.json.simple.JSONObject;
+
+public class EventDataManager extends DatabaseManager {
+
+    /**
+     * Helper function to search JSON arrays
+     */
+    private JSONObject searchJSONArray(JSONArray jsonArray, String key, Object value) {
+        for (Object obj: jsonArray) {
+            JSONObject jsonObject = (JSONObject) obj;
+            if (jsonObject.get(key) == value) {
+                return jsonObject;
+            }
+        }
+        return null;
+    }
+
+    /**
+     * @return EventData with all data from json converted to variables
+     */
+    public EventData fetchEvent(String key, Object value) {
+        JSONArray eventsData = (JSONArray) super.fullDatabase.get("events");
+        JSONObject eventData = searchJSONArray(eventsData, key, value);
+        assert eventData != null;
+        return eventFactory.createEventData(eventData);
+    }
+
+}
