+++ src/main/java/objects/Event_Encounter.java
@@ -1,0 +1,5 @@
+package objects;
+
+public class Event_Encounter extends Event {
+
+}
