+++ src/main/java/objects/Event_ItemPickUp.java
@@ -1,0 +1,5 @@
+package objects;
+
+public class Event_ItemPickUp extends Event {
+
+}
